+++ Projeto2020_API/Controllers/UsuarioController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using Projeto2025_API.Validation;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -9,17 +10,28 @@ namespace Projeto2025_API.Controllers
     public class UsuarioController : ControllerBase
     {
         private readonly IUsuarioService service;
+        private readonly UsuarioValidation validacao;
 
         public UsuarioController(IUsuarioService service)
         {
             this.service = service;
+            this.validacao = new UsuarioValidation();
         }
 
         [HttpPost]
         public async Task<ActionResult<UsuarioDTO>> AddAsync(UsuarioDTO usuarioDTO)
         {
-            var dto = await service.AddAsync(usuarioDTO);
-            return Ok(dto);
+            var result = validacao.Validate(usuarioDTO);
+
+            if (result.IsValid)
+            {
+                var dto = await service.AddAsync(usuarioDTO);
+                return Ok(dto);
+            }
+            else
+            {
+                return BadRequest(result.Errors);
+            }
         }
 
         [HttpGet]
@@ -32,22 +44,38 @@ namespace Projeto2025_API.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<UsuarioDTO>> GetAsync(int id)
         {
+            if (id <= 0)
+                return BadRequest("Id inválido.");
+
             var usuario = await service.GetAsync(id);
             if (usuario == null)
                 return NotFound();
+
             return Ok(usuario);
         }
 
         [HttpPut]
         public async Task<ActionResult> UpdateAsync(UsuarioDTO usuarioDTO)
         {
-            await service.UpdateAsync(usuarioDTO);
-            return NoContent();
+            var result = validacao.Validate(usuarioDTO);
+
+            if (result.IsValid)
+            {
+                await service.UpdateAsync(usuarioDTO);
+                return NoContent();
+            }
+            else
+            {
+                return BadRequest(result.Errors);
+            }
         }
 
         [HttpDelete("{id}")]
         public async Task<ActionResult> RemoveAsync(int id)
         {
+            if (id <= 0)
+                return BadRequest("Id inválido.");
+
             await service.RemoveAsync(id);
             return NoContent();
         }
